// src/types/hero-content.ts
export interface HeroContentConfig {
  id?: string;
  heading: string;
  paragraph: string;
  headingFontSizeDesktop: number;
  headingFontSizeMobile: number;
  paragraphFontSizeDesktop: number;
  paragraphFontSizeMobile: number;
  glassWidthDesktop: string;
  glassHeightDesktop: string;
  glassWidthMobile: string;
  glassHeightMobile: string;
}